Update cs2.json
After the recent changes where valve allowed anonymous login, using an account or simply leaving the login variables blank would cause the installer to fail and then kick pterodactyl out of install mode. The container would restart and carry on the installation in front of the end user as usual, but any additional scripts wouldn't be run. 

This simply makes the installation script use the default anonymous variables and cleans up the login/beta variables that are no longer required.