Update cs2.json
Changed the default IP to use 0.0.0.0 - the same as the original CSGO egg. The reason for this is it will allow for RCON connectivity while in bridge mode. 

I've also added a variable for the GSLT as this is a core function of the game, and it should be there by default.